+++ lib/iex/endpoints/earnings.rb
@@ -2,7 +2,7 @@ module IEX
   module Endpoints
     module Earnings
       def earnings(symbol, options = {})
-        get("stock/#{symbol}/earnings", options)['earnings'].map do |data|
+        get("stock/#{symbol}/earnings", { token: publishable_token }.merge(options))['earnings'].map do |data|
           IEX::Resources::Earnings.new(data)
         end
       rescue Faraday::ResourceNotFound => e